x64: Refactor to remove fake interfaces and general cleanups.
