--- src/common/cpu_detect_generic.cpp
@@ -1,19 +1,0 @@
-// Copyright 2014 Dolphin Emulator Project
-// Licensed under GPLv2+
-// Refer to the license.txt file included.
-
-#include "cpu_detect.h"
-#include "hash.h"
-
-namespace Common {
-
-CPUInfo cpu_info;
-
-CPUInfo::CPUInfo() {
-}
-
-std::string CPUInfo::Summarize() {
-    return "Generic";
-}
-
-} // namespace Common
