+++ src/common/x64/abi.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "common_types.h"
+#include "common/common_types.h"
 
 // x86/x64 ABI:s, and helpers to help follow them when JIT-ing code.
 // All convensions return values in EAX (+ possibly EDX).
@@ -55,7 +55,7 @@
 // 32-bit bog standard cdecl, shared between linux and windows
 // MacOSX 32-bit is same as System V with a few exceptions that we probably don't care much about.
 
-#elif _M_X86_64 // 64 bit calling convention
+#elif ARCHITECTURE_X64 // 64 bit calling convention
 
 #ifdef _WIN32 // 64-bit Windows - the really exotic calling convention
