--- src/video_core/shader/shader_jit_fake.cpp
@@ -1,91 +1,0 @@
-// Copyright 2015 Citra Emulator Project
-// Licensed under GPLv2 or any later version
-// Refer to the license.txt file included.
-
-#include "common/fake_emitter.h"
-
-#include "video_core/shader/shader.h"
-#include "video_core/shader/shader_jit.h"
-
-namespace Pica {
-
-namespace Shader {
-
-using namespace FakeGen;
-
-void Jit::Comp_ADD(Instruction instr) {
-}
-
-void Jit::Comp_DP3(Instruction instr) {
-}
-
-void Jit::Comp_DP4(Instruction instr) {
-}
-
-void Jit::Comp_MUL(Instruction instr) {
-}
-
-void Jit::Comp_FLR(Instruction instr) {
-}
-
-void Jit::Comp_MAX(Instruction instr) {
-}
-
-void Jit::Comp_MIN(Instruction instr) {
-}
-
-void Jit::Comp_MOVA(Instruction instr) {
-}
-
-void Jit::Comp_MOV(Instruction instr) {
-}
-
-void Jit::Comp_SLTI(Instruction instr) {
-}
-
-void Jit::Comp_RCP(Instruction instr) {
-}
-
-void Jit::Comp_RSQ(Instruction instr) {
-}
-
-void Jit::Comp_NOP(Instruction instr) {
-}
-
-void Jit::Comp_END(Instruction instr) {
-}
-
-void Jit::Comp_CALL(Instruction instr) {
-}
-
-void Jit::Comp_CALLC(Instruction instr) {
-}
-
-void Jit::Comp_CALLU(Instruction instr) {
-}
-
-void Jit::Comp_CMP(Instruction instr) {
-}
-
-void Jit::Comp_MAD(Instruction instr) {
-}
-
-void Jit::Comp_IF(Instruction instr) {
-}
-
-void Jit::Comp_LOOP(Instruction instr) {
-}
-
-void Jit::Comp_JMP(Instruction instr) {
-}
-
-void Jit::Comp_NextInstr(unsigned* offset) {
-}
-
-CompiledShader Jit::Compile() {
-    return nullptr;
-}
-
-} // namespace Shader
-
-} // namespace Pica
